Move test of local file system handler
