+++ tests/qgis/test_handler.py
@@ -2,42 +2,9 @@ from qgis.testing import start_app, unittest
 
 from qgis_resource_sharing.repository_handler import (
     BitBucketHandler,
-    FileSystemHandler,
     GithubHandler,
     GogsHandler,
 )
-from tests.qgis.utilities import test_data_path, test_repository_url
-
-
-class TestFileSystemHandler(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        start_app()
-
-    def setUp(self):
-        self.fs_handler = FileSystemHandler(test_repository_url())
-
-    def test_can_handle(self):
-        """Test can_handle function."""
-        url = "file:///home/akbar/dev"
-        self.fs_handler.url = url
-        self.assertTrue(self.fs_handler.can_handle())
-
-        url = "http:///home/akbar/dev"
-        self.fs_handler.url = url
-        self.assertFalse(self.fs_handler.can_handle())
-
-    def test_fetch_metadata(self):
-        """Test fetch_metadata function."""
-        # TC1: Normal successful testcase
-        self.assertIsNone(self.fs_handler.metadata)
-        status, _ = self.fs_handler.fetch_metadata()
-        self.assertTrue(status)
-        self.assertIsNotNone(self.fs_handler.metadata)
-        metadata_path = test_data_path("metadata.ini")
-        with open(metadata_path, "r") as metadata_file:
-            expected_content = metadata_file.read()
-        self.assertEqual(self.fs_handler.metadata, expected_content)
 
 
 class TestRemoteGitHandler(unittest.TestCase):
